ban phase in old draft is now a separate component
