+++ client/src/SessionView/Overview/Drafting/Drafting/index.js
@@ -1,0 +1,1 @@
+export { Drafting } from './Drafting'
